+++ gptables/test/test_gptable.py
@@ -1,3 +1,4 @@
+import numpy as np
 import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
@@ -8,7 +9,9 @@ from gptables import GPTable
 
 
 # TODO: These should be stored in GPTable
-gptable_text_attrs = ["title", "scope", "source"]
+gptable_compulsory_text_attrs = ["title", "instructions"]
+
+gptable_optional_text_attrs = ["scope", "source"]
 
 gptable_list_text_attrs = ["subtitles", "legend"]
 
@@ -20,13 +23,15 @@ valid_index_columns = [
     {1: "one", 2: "two", 3: "three"}
     ]
 
-valid_text_elements = [
+valid_text_elements_excl_none = [
     "This is a string",
     ["This is ", {"bold": True}, "rich", "text"],
-    None
 ]
 
-invalid_text_elements = [
+valid_text_elements_incl_none = valid_text_elements_excl_none.copy()
+valid_text_elements_incl_none.append(None)
+
+invalid_text_elements_excl_none = [
     dict(),
     set(),
     42,
@@ -34,6 +39,8 @@ invalid_text_elements = [
     True
 ]
 
+invalid_text_elements_incl_none = invalid_text_elements_excl_none.copy()
+invalid_text_elements_incl_none.append(None)
 
 @contextmanager
 def does_not_raise():
@@ -48,8 +55,6 @@ def create_gptable_with_kwargs():
             "table": pd.DataFrame(),
             "table_name": "",
             "title": "",
-            "scope": "",
-            "source": "",
             "index_columns": {}  # Override default, as no columns in table
             }
         if format_dict is not None:
@@ -67,25 +72,29 @@ def test_init_defaults(create_gptable_with_kwargs):
 
     # Required args
     assert empty_gptable.title == ""
-    assert empty_gptable.scope == ""
-    assert empty_gptable.units == None
-    assert empty_gptable.source == ""
+    assert empty_gptable.table_name == ""
+
     assert_frame_equal(
             empty_gptable.table, pd.DataFrame().reset_index(drop=True)
             )
-    
+
     # Optional args
+    assert empty_gptable.scope == None
+    assert empty_gptable.source == None
+    assert empty_gptable.units == None
     assert empty_gptable.index_columns == {}
     assert empty_gptable.subtitles == []
     assert empty_gptable.legend == []
-    assert empty_gptable._annotations == []
     assert empty_gptable.additional_formatting == []
-    
+    assert empty_gptable.instructions == "This worksheet contains one table. Some cells may refer to notes, which can be found on the notes worksheet."
+
     # Other
     assert empty_gptable.index_levels == 0
     assert empty_gptable._column_headings == set()
     assert empty_gptable._VALID_INDEX_LEVELS == [1, 2, 3]
-    
+    assert empty_gptable._annotations == []
+    assert empty_gptable.data_range == [2, 0, 2, -1]
+
 
 class TestAttrValidationGPTable:
     @pytest.mark.parametrize("level", [4, 0, -1])
@@ -151,11 +160,58 @@ class TestAttrValidationGPTable:
         """
         with pytest.raises(TypeError):
             create_gptable_with_kwargs({"table": not_a_table})
-    
 
-    @pytest.mark.parametrize("attr", gptable_text_attrs)
-    @pytest.mark.parametrize("not_text", invalid_text_elements)
-    def test_invalid_text_in_str_attrs(self, attr, not_text, create_gptable_with_kwargs):
+
+    def test_set_table_name(self, create_gptable_with_kwargs):
+        """
+        Test that setting GPTable table name with a valid string works as expected
+        """
+        gptable = create_gptable_with_kwargs({"table_name": "table_name"})
+        assert gptable.table_name == "table_name"
+
+
+    @pytest.mark.parametrize("invalid_name", invalid_text_elements_incl_none)
+    def test_invalid_type_table_name(self, invalid_name, create_gptable_with_kwargs):
+        """
+        Test that setting GPTable table name to object other than string raises
+        an error.
+        """
+        with pytest.raises(TypeError):
+            create_gptable_with_kwargs({"table_name": invalid_name})
+
+
+    def test_table_name_not_list(self, create_gptable_with_kwargs):
+        """
+        Test that setting GPTable table name to a list, eg with rich text,
+        raises an error.
+        """
+        with pytest.raises(TypeError):
+            create_gptable_with_kwargs({"table_name": []})
+
+
+    @pytest.mark.parametrize("invalid_name", [" ", "  "])
+    def test_invalid_characters_table_name(self, invalid_name, create_gptable_with_kwargs):
+        """
+        Test that setting GPTable table name to string with whitespace raises
+        an error.
+        """
+        with pytest.raises(ValueError):
+            create_gptable_with_kwargs({"table_name": invalid_name})
+
+
+    @pytest.mark.parametrize("attr", gptable_compulsory_text_attrs)
+    @pytest.mark.parametrize("not_text", invalid_text_elements_incl_none)
+    def test_invalid_text_in_compulsory_str_attrs(self, attr, not_text, create_gptable_with_kwargs):
+        """
+        Test that setting an invalid GPTable text types raises a TypeError for
+        each attribute that holds a string.
+        """
+        with pytest.raises(TypeError):
+            create_gptable_with_kwargs({attr: not_text})
+
+    @pytest.mark.parametrize("attr", gptable_optional_text_attrs)
+    @pytest.mark.parametrize("not_text", invalid_text_elements_excl_none)
+    def test_invalid_text_in_optional_str_attrs(self, attr, not_text, create_gptable_with_kwargs):
         """
         Test that setting an invalid GPTable text types raises a TypeError for
         each attribute that holds a string.
@@ -164,9 +220,24 @@ class TestAttrValidationGPTable:
             create_gptable_with_kwargs({attr: not_text})
     
 
-    @pytest.mark.parametrize("attr", gptable_text_attrs)
-    @pytest.mark.parametrize("text", valid_text_elements)
-    def test_valid_text_in_str_attrs(self, attr, text, create_gptable_with_kwargs):
+    @pytest.mark.parametrize("attr", gptable_compulsory_text_attrs)
+    @pytest.mark.parametrize("text", valid_text_elements_excl_none)
+    def test_valid_text_in_compulsory_str_attrs(self, attr, text, create_gptable_with_kwargs):
+        """
+        Test that setting valid GPTable text elements works as expected. Test
+        strings and list containing strings and format dicts (rich text).
+        Also test that None is allowed.
+        """
+        gptable = create_gptable_with_kwargs({attr: text})
+
+        if isinstance(text, list):
+            assert getattr(gptable, attr).list == text
+        else:
+            assert getattr(gptable, attr) == text
+
+    @pytest.mark.parametrize("attr", gptable_optional_text_attrs)
+    @pytest.mark.parametrize("text", valid_text_elements_incl_none)
+    def test_valid_text_in_compulsory_str_attrs(self, attr, text, create_gptable_with_kwargs):
         """
         Test that setting valid GPTable text elements works as expected. Test
         strings and list containing strings and format dicts (rich text).
@@ -181,7 +252,7 @@ class TestAttrValidationGPTable:
 
 
     @pytest.mark.parametrize("attr", gptable_list_text_attrs)
-    @pytest.mark.parametrize("text", invalid_text_elements)
+    @pytest.mark.parametrize("text", invalid_text_elements_incl_none)
     def test_invalid_text_in_list_attrs(self, attr, text, create_gptable_with_kwargs):
         """
         Test that setting list of invalid text elements to GPTable list
@@ -193,28 +264,22 @@ class TestAttrValidationGPTable:
 
 
     @pytest.mark.parametrize("attr", gptable_list_text_attrs)
-    @pytest.mark.parametrize("text", valid_text_elements)
+    @pytest.mark.parametrize("text", valid_text_elements_excl_none)
     def test_valid_text_in_list_attrs(self, attr, text, create_gptable_with_kwargs):
         """
         Test that setting list of valid text elements to GPTable list
         parameters works as expected.
         """
-        if text is not None:
-            text_list = [text, text]
-        else:
-            text_list = []
-
+        text_list = [text, text]
         gptable = create_gptable_with_kwargs({attr: text_list})
 
-        if isinstance(text, str):
-            assert getattr(gptable, attr) == text_list
-        elif isinstance(text, list):
+        if isinstance(text, list):
             assert all([element.list == text for element in getattr(gptable, attr)])
         else:
-            assert getattr(gptable, attr) == []
-        
+            assert getattr(gptable, attr) == text_list
 
-    @pytest.mark.parametrize("key", invalid_text_elements[2:] + ["invalid_key"])
+
+    @pytest.mark.parametrize("key", invalid_text_elements_incl_none[2:] + ["invalid_key"])
     def test_invalid_additional_format_keys(self, key, create_gptable_with_kwargs):
         """
         Test that adding additional formatting with an invalid key raises an
@@ -239,7 +304,7 @@ class TestAttrValidationGPTable:
         assert getattr(gptable, "additional_formatting") == [{key: {"format": {"bold": True}}}]
     
 
-    @pytest.mark.parametrize("format_label", invalid_text_elements[2:] + ["not_a_format"])
+    @pytest.mark.parametrize("format_label", invalid_text_elements_incl_none[2:] + ["not_a_format"])
     def test_invalid_additional_format_labels(self, format_label, create_gptable_with_kwargs):
         """
         Test that adding additional formatting with a format parameter that is
@@ -284,13 +349,129 @@ class TestAttrValidationGPTable:
         assert getattr(gptable, "additional_formatting") == additional_formatting
 
 
-    @pytest.mark.parametrize("column_names,expectation", [
-        (["columnA", "columnB", "columnC"], does_not_raise()),
-        (["columnA", "columnB", ""], pytest.raises(ValueError))
-    ])
-    def test_validate_all_column_names_have_text(self, column_names, expectation, create_gptable_with_kwargs):
+    @pytest.mark.parametrize("unit_text", valid_text_elements_excl_none)
+    @pytest.mark.parametrize("column_id", ["columnA", 0])
+    def test_units_placement(self, unit_text, column_id, create_gptable_with_kwargs):
         """
-        Test that GPTable raises error when there are empty strings for column names.
+        Test that units are placed correctly under column headers.
+        """
+        table_with_units = pd.DataFrame(columns=[f"columnA\n({unit_text})"])
+
+        gptable = create_gptable_with_kwargs({
+            "table": pd.DataFrame(columns=["columnA"]),
+        })
+
+        gptable.set_units(new_units = {column_id: unit_text})
+
+        assert gptable.table.columns == table_with_units.columns
+
+
+    @pytest.mark.parametrize("column_id", ["columnA", 0])
+    def test_table_notes_placement(self, column_id, create_gptable_with_kwargs):
+        """
+        Test that units are placed correctly under column headers.
+        """
+        table_with_notes = pd.DataFrame(columns=[f"columnA\n$$note_reference$$"])
+
+        gptable = create_gptable_with_kwargs({
+            "table": pd.DataFrame(columns=["columnA"])
+        })
+
+        gptable.set_table_notes(
+            new_table_notes = {column_id: "$$note_reference$$"}
+        )
+
+        assert gptable.table.columns == table_with_notes.columns
+
+
+    @pytest.mark.parametrize("column_id", ["columnA", 0])
+    def test_table_units_and_notes_placement(self, column_id, create_gptable_with_kwargs):
+        """
+        Test that units and notes are placed correctly under column headers.
+        """
+        table_with_units_and_notes = pd.DataFrame(
+            columns=[f"columnA\n(unit)\n$$note_reference$$"]
+        )
+
+        gptable = create_gptable_with_kwargs({
+            "table": pd.DataFrame(columns=["columnA"])
+        })
+
+        gptable.set_units(new_units = {column_id: "unit"})
+        gptable.set_table_notes(new_table_notes = {column_id: "$$note_reference$$"})
+
+        assert gptable.table.columns == table_with_units_and_notes.columns
+
+
+    def test_additional_formatting_with_units(self, create_gptable_with_kwargs):
+        """
+        Test that units are placed correctly under column headers.
+        """
+
+        gptable = create_gptable_with_kwargs({
+            "table": pd.DataFrame(columns=["columnA"]),
+            "units": {"columnA": "unit"},
+            "additional_formatting": [{"column": {
+                    "columns": ["columnA"],
+                    "format": {"bold": True}
+                    }}]
+        })
+
+        assert gptable.additional_formatting == [{"column": {
+            "columns": ["columnA\n(unit)"],
+            "format": {"bold": True}
+        }}]
+
+
+    def test_additional_formatting_with_table_notes(self, create_gptable_with_kwargs):
+        """
+        Test that units are placed correctly under column headers.
+        """
+
+        gptable = create_gptable_with_kwargs({
+            "table": pd.DataFrame(columns=["columnA"]),
+            "table_notes": {"columnA": "$$ref$$"},
+            "additional_formatting": [{"column": {
+                    "columns": ["columnA"],
+                    "format": {"bold": True}
+                    }}]
+        })
+
+        assert gptable.additional_formatting == [{"column": {
+            "columns": ["columnA\n$$ref$$"],
+            "format": {"bold": True}
+        }}]
+
+
+    def test_additional_formatting_with_units_and_table_notes(self, create_gptable_with_kwargs):
+        """
+        Test that units are placed correctly under column headers.
+        """
+
+        gptable = create_gptable_with_kwargs({
+            "table": pd.DataFrame(columns=["columnA"]),
+            "units": {"columnA": "unit"},
+            "table_notes": {"columnA": "$$ref$$"},
+            "additional_formatting": [{"column": {
+                    "columns": ["columnA"],
+                    "format": {"bold": True}
+                    }}]
+        })
+
+        assert gptable.additional_formatting == [{"column": {
+            "columns": ["columnA\n(unit)\n$$ref$$"],
+            "format": {"bold": True}
+        }}]
+
+
+    @pytest.mark.parametrize("column_names,expectation", [
+        (["columnA", "columnB"], does_not_raise()),
+        (["columnA", ""], pytest.raises(ValueError)),
+        ([None, "columnB"], pytest.raises(ValueError))
+    ])
+    def test__validate_all_column_names_have_text(self, column_names, expectation, create_gptable_with_kwargs):
+        """
+        Test that GPTable raises error when there are null values or empty strings for column names.
         """
         with expectation:
             create_gptable_with_kwargs({
@@ -302,7 +483,7 @@ class TestAttrValidationGPTable:
         (["columnA", "columnB", "columnC"], does_not_raise()),
         (["columnA", "columnB", "columnB"], pytest.raises(ValueError))
     ])
-    def test_validate_no_duplicate_column_names(self, column_names, expectation, create_gptable_with_kwargs):
+    def test__validate_no_duplicate_column_names(self, column_names, expectation, create_gptable_with_kwargs):
         """
         Test that GPTable raises error when there are duplicate column names in table data.
         """
@@ -313,11 +494,11 @@ class TestAttrValidationGPTable:
 
 
 
-@pytest.mark.parametrize("index_cols", valid_index_columns)
 class TestIndirectAttrs:
     """
     Test that non-formatting create_gptable_with_kwargs attributes are indirectly set correctly.
     """
+    @pytest.mark.parametrize("index_cols", valid_index_columns)
     def test_index_levels_set(self, index_cols, create_gptable_with_kwargs):
         """
         Test that number of index levels are set, when one, two or three
@@ -334,8 +515,9 @@ class TestIndirectAttrs:
             "index_columns": index_cols
             })
         assert getattr(gptable, "index_levels") == len(index_cols)
-        
 
+
+    @pytest.mark.parametrize("index_cols", valid_index_columns)
     def test_column_headings_set(self, index_cols, create_gptable_with_kwargs):
         """
         Test that non-index columns are set as column headings.
@@ -355,3 +537,30 @@ class TestIndirectAttrs:
         exp = set(range(4)) - set(range(len(index_cols)))
         
         assert getattr(gptable, "_column_headings") == exp
+
+
+    def test__annotations_set(self, create_gptable_with_kwargs):
+        """
+        Test that annotation references in `gptable` attributes are found and
+        added to `_annotations` as expected.
+        """
+        table = pd.DataFrame(columns=["col"])
+
+        kwargs = {
+            "title": "Title$$1$$",
+            "subtitles": ["Subtitle$$2$$"],
+            "instructions": "Instructions$$3$$",
+            "source": "Source$$4$$",
+            "scope": "Scope$$5$$",
+            "legend": ["Legend$$6$$"],
+            "units": {0: "Unit$$7$$"},
+            "table_notes": {0: "Note$$8$$"},
+            "table": table
+        }
+
+        gptable = create_gptable_with_kwargs(kwargs)
+
+        description_order = ["instructions", "source", "scope", "legend"]
+        gptable._set_annotations(description_order)
+
+        assert gptable._annotations == ["1", "2", "3", "4", "5", "6", "7", "8"]